תיקונים בשביל הmakefile
שיניתי סדר של דברים ועוד כל מיני שטויות כדי שזה יתקמפל בלי אזהרות שלא יזיינו תמוח

+++ Trainer.cpp
@@ -77,7 +77,7 @@ void Trainer::closeTrainer() {
 }
 
 bool Trainer::canAcceptAnotherCustomer() {
-    return customersList.size() < capacity;
+    return (int)customersList.size() < capacity;
 }
 
 Trainer::~Trainer() {
